added Auth token from github for private directory 3

+++ server.js
@@ -10,25 +10,29 @@ const port = 80;
 // Middleware to parse JSON bodies
 app.use(bodyParser.json());
 
-// Endpoint to handle webhook POST requests
 app.post('/webhook', (req, res) => {
-  const command = `
-    cd ./project_directory &&
-    git config --global credential.username rphllc &&
-    echo "https://${process.env.GITHUB_USERNAME}:${process.env.GITHUB_TOKEN}@github.com" > ~/.git-credentials &&
-    git pull &&
-    docker-compose down &&
-    docker-compose up --build -d
-    `;
-
-  exec(command, (error, stdout, stderr) => {
-    if (error) {
-      console.error(`exec error: ${error}`);
-      return res.status(500).send('Server Error');
+  const setupCredentials = `git config --global credential.helper 'store' && \\
+                            echo "https://${process.env.GITHUB_USERNAME}:${process.env.GITHUB_TOKEN}@github.com" > ~/.git-credentials`;
+  exec(setupCredentials, (setupError) => {
+    if (setupError) {
+      console.error(`Setup error: ${setupError}`);
+      return res.status(500).send('Failed to setup Git credentials');
     }
-    console.log(`stdout: ${stdout}`);
-    console.error(`stderr: ${stderr}`);
-    res.status(200).send('Docker update triggered');
+    const command = `
+          cd ./project_directory &&
+          git pull &&
+          docker-compose down &&
+          docker-compose up --build -d
+      `;
+    exec(command, (error, stdout, stderr) => {
+      if (error) {
+        console.error(`Exec error: ${error}`);
+        return res.status(500).send('Server Error');
+      }
+      console.log(`stdout: ${stdout}`);
+      console.error(`stderr: ${stderr}`);
+      res.status(200).send('Docker update triggered');
+    });
   });
 });
